+++ acme/protocols/HttpServer.py
@@ -32,7 +32,7 @@ from ..etc.Utils import renameThread, getThreadName, isURL, getBasicAuthFromUrl
 from ..helpers.TextTools import findXPath
 from ..helpers.MultiDict import MultiDict
 from ..etc.DateUtils import timeUntilAbsRelTimestamp, getResourceDate, rfc1123Date
-from ..etc.RequestUtils import toHttpUrl, serializeData, deserializeData, requestFromResult
+from ..etc.RequestUtils import toHttpUrl, serializeData, deserializeData, requestFromResult, prepareResultForSending
 from ..etc.RequestUtils import createPositiveResponseResult, fromHttpURL, contentAsString, fillRequestWithArguments
 from ..helpers.NetworkTools import isTCPPortAvailable
 from ..runtime.Configuration import Configuration
@@ -420,7 +420,7 @@ class HttpServer(object):
 		return Response(response = 'unsupported', status = 422, headers = self._responseHeaders)
 
 
-	def handleUpperTester(self, path:Optional[str] = None) -> Response:
+	def handleUpperTester(self, path:Optional[str]=None) -> Response:
 		"""	Handle a Upper Tester request. See TS-0019 for details.
 		"""
 		if self.isStopped:
@@ -431,39 +431,64 @@ class HttpServer(object):
 			return Response(status = 401)
 
 
-		def prepareUTResponse(rcs:ResponseStatusCode, result:str) -> Response:
+		def prepareUTResponse(rcs:ResponseStatusCode, result:Optional[str]=None, body:Optional[str|bytes]=None) -> Response:
 			"""	Prepare the Upper Tester Response.
+
+				Args:
+					rcs: The response status code.
+					result: The result to be returned.
+
+				Return:
+					The response object.
 			"""
 			headers = {}
 			headers['Server'] = self.serverID
 			headers['X-M2M-RSC'] = str(rcs.value)	# Set the ResponseStatusCode accordingly
 			if result:								# Return an optional return value
 				headers['X-M2M-UTRSP'] = result
-			resp = Response(status = 200 if rcs == ResponseStatusCode.OK else 400, headers = headers)
+			resp = Response(status=200 if rcs == ResponseStatusCode.OK else 400, headers=headers, response=body)
 			L.isDebug and L.logDebug(f'<== Upper Tester Response:') 
 			L.isDebug and L.logDebug(f'Headers: \n{str(resp.headers).rstrip()}')
 			return resp
 
+
 		L.enableScreenLogging and renameThread('UT')
 		L.isDebug and L.logDebug(f'==> Upper Tester Request:') 
 		L.isDebug and L.logDebug(f'Headers: \n{str(request.headers).rstrip()}')
-		if request.data:
-			L.isDebug and L.logDebug(f'Body: \n{request.json}')
 
 		# Handle special commands
 		if (cmd := request.headers.get('X-M2M-UTCMD')) is not None:
 			cmd, _, arg = cmd.partition(' ')
-			if not (res := CSE.script.run(cmd, arg, metaFilter = [ 'uppertester' ], ignoreCase = True))[0]:
+			if not (res := CSE.script.run(cmd, arg, metaFilter=[ 'uppertester' ], ignoreCase=True))[0]:
 				return prepareUTResponse(ResponseStatusCode.BAD_REQUEST, str(res[1]))
 			
 			if res[1].type in [SType.tList, SType.tListQuote]:
 				_r = ','.join(res[1].raw())
 			else:
-				_r = res[1].toString(quoteStrings = False, pythonList = True)
+				_r = res[1].toString(quoteStrings=False, pythonList=True)
 			return prepareUTResponse(ResponseStatusCode.OK, _r)
+		
+		# Treat the request as a normal UT request
+		
+		# Extract the request from the body
+		L.isDebug and L.logDebug(f'Body: \n{request.data!r}')
+		if request.data:
+			try:
+				# Dissect the request
+				dissectResult = CSE.request.dissectRequestFromBytes(request.data, ContentSerializationType.getType(request.content_type))
+				# Directly handle the request
+				responseResult = CSE.request.handleRequest(dissectResult.request)
+			except ResponseException as e:
+				return prepareUTResponse(ResponseStatusCode.BAD_REQUEST, body=f'{{ "m2m:dbg" : "{e.dbg}" }}')
+			
+			# Prepare and send the response
+			_rs, _b = prepareResultForSending(responseResult.prepareResultFromRequest(dissectResult.request),
+									 		  True, 
+											  dissectResult.request)
+			return prepareUTResponse(ResponseStatusCode.OK, body=_b)
 
-		L.logWarn('UT functionality is not fully supported.')
-		return prepareUTResponse(ResponseStatusCode.BAD_REQUEST, None)
+		# Return an error if no body or command is present
+		return prepareUTResponse(ResponseStatusCode.BAD_REQUEST, L.logWarn('UT requires request body or X-M2M-UTCMD.'))
 
 
 	#########################################################################